Add New Protocols and Basic ItemDetailsViewController
MainViewController is a new protocol that makes it easier to ensure that all main UIViewControllers have their expected methods and properties. Similarly, the MainViewModel protocol does the same for each MainViewController's view model. Also a basic implementation of ItemDetailsViewController that's shown when a row is tapped in HomeViewController has been added.