Add Basic NFC Tag Implementation, notes and hasTag Properties for Item
NfcService uses various methods to read and write data from and to NFC tags. At this time, all reading an NFC tag does is print its data to the console, and new data is temporarily no being written to the database. Instead, the user is being prompted to state whether or not they want to use an NFC tag for a new item. Also, Item now has new hasTag and notes properties. These properties are used to identify when an Item has an NFC tag associated with it and any notes about a given Item, respectively.