Add Bug Fixes and Documentation
nfcSessionFinished notification now passes the item that was created or updated in its userInfo dictionary
Fixed bugs in addItemToItemsArray and updateItemInItemsArray methods in HomeViewModel. The update method is called when the user scans an item into a tag. This is necessary because the items in the HomeViewModel items dictionary need to have their hasTag property set to true. Also the add method now sorts the items array again after the new item is appended