Add Ship Item Functionality
It is now possible to delete an item via ItemDetailsViewController. When an item is shipped, an alert appears telling the user to delete the item's data from an NFC tag before they ship the item. In the future, this functionality will be smarter, as this alert should not show when the item being shipped has no tag. They can then follow the prompt in the alert to delete the item from the tag. Once that's finished, the item is automatically deleted from the database.