Add Bug Fixes for Item Updating
Fixed a bug that would prevent ItemDetailsViewController from being updated after an item was updated. This was done with a new .itemUpdated Notification. Another bug was fixed where editing an item and not making any changes before saving would cause some properties to save with empty values. This was happening because the AddEditItemViewModel init was not assigning the passed in itemToEdit to the properties that the updateItem(_:) method was using to create an updated item.