Style AddEditItemTextViewCell and Fix Spacing Bug in AddEditItemViewController
Made the space even between all cells in AddEditItemViewController. The fields were being centered in their cells, which was causing inconsistent spacing because the height of the UITextView cell is higher than that of the UITextField cells. Additionally, the UITextView in AddEditItemViewController now has a Notes placeholder to match the styling of the rest of the fields in the ViewController.