Add Bug Fix for UITextView in AddEditItemViewController
Fixed a bug that was causing the UITextView in AddEditItemViewController to get emptied after starting to scan for an NFC Tag. This was occurring because, when scanning begins, the UITextView's textViewDidBeginEditing gets called. This method would empty the UITextView's text when creating a new item because it was not checking for existing text that the user typed in viewModel.itemNotes.