Add Full Dark Mode Support
Added a custom color to the Assets Catalog for the UITextView in AddEditItemView. This color is black in dark mode and white in light mode. This was necessary instead of using UIColor.systemBackground because UIColor.systemBackground does not set black as the background color on a UITextView in dark mode. Instead, it's a dark gray.